// wwwroot/app/enums/event-type.ts
export enum EventType {
    LatePickUp = 1,
    LateDelivery = 2,
    EtaRequest = 3,
    PodRequest = 4,
    JobChange = 5,
    CancelJob = 6,
    Compliment = 7,
    DrivingComplaint = 8,
    SuburbWrong = 9,
    JobTypeWrong = 10,
    SpeedWrong = 11,
    ContactWrong = 12,
    CustomerWrong = 13,
    DetailsMissing = 14,
    SequenceWrong = 15,
    LateCallMissed = 16,
    LateDelMissed = 17,
    VehicleStock = 18,
    RtProcedure = 19,
    LateStart = 20,
    Breakdown = 21,
    CourierGeneral = 22,
    WrongDeliveryPoint = 39,
    IncorrectPod = 24,
    SizeWrong = 25,
    WrongPickUpPoint = 26,
    PrebookWrong = 27,
    LateCallOn = 28,
    PossibleDataError = 29,
    JobIdleOver1Min = 30,
    DataUpdate = 31,
    ChangeCustomer = 32,
    RestoreJob = 33,
    PriceWrong = 34,
    ChangePrice = 35,
    PrebookVoid = 37,
    InvalidJob = 36,
    CustomerMistake = 38,
    CourierServiceRequired = 40,
    OtherWrong = 41,
    DirectAsap = 42,
    RefAChanged = 43,
    RefBChanged = 44,
    OurRefChanged = 45,
    NotifiedChanged = 46,
    CallMePlease = 47,
    Closed = 48,
    AmRedelRequest = 49,
    ChocolatePromo = 68,
    Callout = 51,
    AddressIncorrect = 52,
    DeliveryAddrIncorrectPleaseLookup = 53,
    FlightDetails = 54,
    DangerousGoods = 69,
    PermissionToLeaveItemAt = 56,
    Gps = 57,
    CallingCard = 70,
    MailNoMail = 71,
    WaitingForJob = 60,
    RestartJob = 61,
    CarNeeded = 62,
    VanNeeded = 63,
    TruckNeeded = 64,
    SuburbChangedTo = 65,
    Other = 66,
    CancelMySupport = 67,
    JobRestarted = 72,
    WebUpdated = 73,
    UpsetClient = 74,
    AddressIncorrectPickup = 75,
    AddressIncorrectDelivery = 76,
    CallingCardContactedClient = 77,
    CallingCardHaveNotContactedClient = 78,
    CalloutCancelled = 79,
    CalloutRebookForLater = 80,
    CalloutSwapPickupAndDeliveryAddress = 81,
    CalloutOther = 82,
    ClosedMorningRedelivery = 83,
    ClosedRedeliverLaterToday = 84,
    DangerousGoodsHaveDocumentation = 85,
    DangerousGoodsNoDocumentation = 86,
    WaitingToCollectJob5Min = 87,
    WaitingToCollectJob10Min = 88,
    WaitingToCollectJob15Min = 89,
    WaitingToCollectJob20Min = 90,
    WaitingToCollectJob30Min = 91,
    Complaint = 92,
    CalloutDeliveryOther = 94,
    CustomerFeedbackHappyCustomer = 95,
    CustomerFeedbackIssueWithCustomer = 96,
    CustomerFeedbackUnderAttack = 97,
    CustomerFeedbackUnhappyCustomer = 98,
    TruckOnSite = 99,
    TruckHandUnload = 100,
    TruckTailLiftPickup = 101,
    TruckTailLiftDelivery = 102,
    TruckOversize = 103,
    TruckOverweight = 104,
    SpeedChanged = 105,
    VehicleChanged = 106,
    QuantityChanged = 107
}